refactor: split everything into modules
