+++ README.md
@@ -1,0 +1,14 @@
+# Messy Kitty
+
+A game where you try to get an old grandma's *out of control cat*
+away from creating a mess.
+
+## Installation
+
+First, you need Python 3. It's probably installed on your system already.
+Then, you need to install *pygame* and the game's dependencies.
+
+Using **pip**:
+```sh
+pip install -r requirements.txt
+```
