+++ game/__init__.py
@@ -1,0 +1,11 @@
+##
+## Epitech Game Jam - May 2025
+## Messy kitty
+##
+## Game module core.
+##
+
+from .game import Game
+
+if __name__ == "__main__":
+    Game().run()
