+++ game/constants.py
@@ -1,0 +1,20 @@
+##
+## Epitech Game Jam - May 2025
+## Messy kitty
+##
+## Default values used within the program
+## (colors, asset paths...)
+##
+
+import pygame as pg
+
+from enum import Enum
+
+class Colors:
+    cyan = pg.Color(0, 255, 255)
+
+class Assets:
+    grandma_path = "assets/grandma.png"
+
+class PossibleEvents(Enum):
+    ...
