+++ game/events.py
@@ -1,0 +1,11 @@
+##
+## Epitech Game Jam - May 2025
+## Messy kitty
+##
+## Event module core.
+##
+
+class Event:
+    def __init__(self):
+        self.event_total: int = 5
+        self.event_prob: float = 1 # In percentage
